Move database functions to separate file.
